Makefile bug fix: Clean bogus use of make variables
The makefile appended to a dirty make variable also used by other makefiles.
As a result, the next project using that variable would rely on an incorrect
list of LOCAL_C_INCLUDES (part irrelevant).
A side effect of that bug was, that the toolkit would fail to pack because
include paths not approved for packing would be required by the next project.

This change cleans the makefile by:
- Renaming all common_... variables to the more unique libseccomp_... names.
- Changing all incorrect uses of += to a plain := assignment
- Fixing LOCAL_PATH and other paths implied by it, to use the make infrastucture
  in a more expected way, as other projects use them

Change-Id: Ibd69404202474b43b4393c3b0a60004427bc4d6f